@@ -2,7 +2,7 @@ import { Component, OnInit, EventEmitter, Output } from '@angular/core';
 import { LoginMenuAction, IMenuItem, LoginMenuActionTypes } from '../../authentication.models';
 import { AppState } from '../../../app.data.models';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'store-login-icon',
