@@ -1,11 +1,14 @@
+
+import {empty as observableEmpty,  Observable } from 'rxjs';
+
+import {switchMap, catchError, map} from 'rxjs/operators';
 import { ApplicationsService } from './applications.service';
 import { Injectable } from '@angular/core';
 import { Actions, Effect } from '@ngrx/effects';
 import * as applicationsActions from './applications.actions';
 import { Application, Subscription } from './applications.data.models';
 import { NotificationService } from '../shared/services/notification.service';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
+
 import { HttpErrorResponse } from '@angular/common/http';
 import { GET_APPLICATION_SUBSCRIPTIONS, GetApplicationSubscriptionsSuccessAction } from './applications.actions';
 
@@ -19,40 +22,40 @@ export class ApplicationsEffects {
 
   @Effect()
   countries$ = this.actions$
-    .ofType(applicationsActions.GET_ALL_APPLICATIONS)
-    .map(
+    .ofType(applicationsActions.GET_ALL_APPLICATIONS).pipe(
+    map(
       (action: applicationsActions.GetAllApplicationsAction) => action.payload
-    )
-    .switchMap(param =>
+    ),
+    switchMap(param =>
       this.service
-        .getAllApplications(param)
-        .map(
+        .getAllApplications(param).pipe(
+        map(
           (result: Application[]) =>
             new applicationsActions.GetAllApplicationsSuccessAction(result)
-        )
-        .catch((e: HttpErrorResponse) => {
+        ),
+        catchError((e: HttpErrorResponse) => {
           this.notification.error(e.message);
-          return Observable.empty();
-        })
-    );
+          return observableEmpty();
+        }),)
+    ),);
 
 
   @Effect()
   appSubscriptions$ = this.actions$
-    .ofType(applicationsActions.GET_APPLICATION_SUBSCRIPTIONS)
-    .map(
+    .ofType(applicationsActions.GET_APPLICATION_SUBSCRIPTIONS).pipe(
+    map(
       (action: applicationsActions.GetApplicationSubscriptionsAction) => action.payload
-    )
-    .switchMap(param =>
+    ),
+    switchMap(param =>
       this.service
-        .getApplicationSubscriptions(param)
-        .map(
+        .getApplicationSubscriptions(param).pipe(
+        map(
           (result: Subscription[]) =>
             new applicationsActions.GetApplicationSubscriptionsSuccessAction(result)
-        )
-        .catch((e: HttpErrorResponse) => {
+        ),
+        catchError((e: HttpErrorResponse) => {
           this.notification.error(e.message);
-          return Observable.empty();
-        })
-    );
+          return observableEmpty();
+        }),)
+    ),);
 }
