@@ -1,6 +1,8 @@
+
+import {filter} from 'rxjs/operators';
 import { Component, OnInit, Input } from '@angular/core';
 import { Router, NavigationEnd } from '@angular/router';
-import 'rxjs/add/operator/filter';
+
 
 @Component({
   selector: 'store-breadcrumbs',
@@ -17,8 +19,8 @@ export class BreadcrumbsComponent implements OnInit {
   }
 
   ngOnInit() {
-    this._router.events
-      .filter((event: any) => event instanceof NavigationEnd)
+    this._router.events.pipe(
+      filter((event: any) => event instanceof NavigationEnd))
       .subscribe((event: NavigationEnd) => {
         const tmp = event.url.replace('/', '').split('/');
         if (tmp[0] === 'legacy') {
